+++ src/main/java/ooga/model/entities/maincharacter/ImmutableMainCharacterBehavior.java
@@ -1,10 +1,8 @@
 package ooga.model.entities.maincharacter;
 
-import ooga.model.MainCharacterState;
-
 public interface ImmutableMainCharacterBehavior {
 
-  double getScore();
+  int getScore();
   MainCharacterState getMainCharacterState();
 
 }
