+++ src/main/java/ooga/model/entities/maincharacter/MainCharacterBehavior.java
@@ -1,13 +1,11 @@
 package ooga.model.entities.maincharacter;
 
-import ooga.model.MainCharacterState;
-
 public interface MainCharacterBehavior extends ImmutableMainCharacterBehavior{
 
   MainCharacterState getMainCharacterState();
   void checkLivesAndUpdateMainCharacterState();
   void setMainCharacterState(MainCharacterState mainCharacterState);
-  void updateScore(double addToScore);
-  double getScore();
+  void updateScore(int addToScore);
+  int getScore();
 
 }
